Summarizer/requirements.txt
streamlit
openai
PyMuPDF
numpy
python-dotenv